requirements.txt
pyspark
flask
psycopg2-binary
pytest
requests
